requirements.txt
ccxt>=4.1.13
python-dotenv>=1.0.0
pandas>=2.1.0
requests>=2.31.0
websockets>=12.0
aiohttp>=3.9.1
